requirements.txt
playwright
requests
gspread
google-auth
